requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.4
Django==4.1.4
django-celery-beat==2.4.0
django-rest-knox==4.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.0.4
greenlet==2.0.1
iniconfig==1.1.1
kombu==5.2.4
packaging==22.0
pluggy==1.0.0
prompt-toolkit==3.0.36
pycparser==2.21
pytest==7.2.0
pytest-django==4.5.2
python-crontab==2.7.0
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.0
six==1.16.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
